% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Aphanopus_EF}
\alias{Aphanopus_EF}
\title{Aphanopus elliptic Fourier descriptors}
\format{
A data frame with 129 variables: `Species`, `ID`, and Fourier coefficients `A2` through `A32`, `B1`..`B32`, `C1`..`C32`, and `D1`..`D32`.
}
\usage{
Aphanopus_EF
}
\description{
Elliptic Fourier descriptors for Aphanopus carbo and A. intermedius otoliths.
}
\keyword{datasets}
