% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Aphanopus_W5}
\alias{Aphanopus_W5}
\title{Aphanopus wavelet coefficients (Scale 5)}
\format{
A data frame with 514 variables: `Species`, `ID`, and wavelet coefficients from `W1` to `W512`.
}
\usage{
Aphanopus_W5
}
\description{
Wavelet coefficients at scale 5 for Aphanopus carbo and A. intermedius otoliths.
}
\keyword{datasets}
