% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{.compute_curvature_raw}
\alias{.compute_curvature_raw}
\title{Internal helper for raw curvature}
\usage{
.compute_curvature_raw(coo)
}
\description{
Internal helper for raw curvature
}
\keyword{internal}
