% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Aphanopus_EF}
\alias{Aphanopus_EF}
\title{Aphanopus elliptic Fourier descriptors}
\format{
A data frame with 130 variables:
\describe{
  \item{Species}{Species name}
  \item{ID}{Unique identifier}
  \item{A1..D32}{Fourier coefficients (harmonincs 1 to 32)}
}
}
\usage{
Aphanopus_EF
}
\description{
Elliptic Fourier descriptors for Aphanopus carbo and A. intermedius otoliths.
}
\keyword{datasets}
