% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatrous1d.R
\name{fatrous1d}
\alias{fatrous1d}
\title{Atrous Convolution}
\usage{
fatrous1d(rx, rh, snmin, snmax, sj)
}
\arguments{
\item{rx}{A numeric vector representing the input signal.}

\item{rh}{A numeric vector representing the filter coefficients.}

\item{snmin}{The minimum scale index.}

\item{snmax}{The maximum scale index.}

\item{sj}{The scale factor (must be >= 0).}
}
\value{
A matrix representing the atrous convolution result.
}
\description{
Computes the atrous convolution of a signal.
}
\examples{
# Example usage:
ra1 <- fatrous1d(ra0, rh_, snh_min, snh_max, 1)
rd1 <- fatrous1d(ra0, rg_, sng_min, sng_max, 1)
}
