% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwaveletspl_3.R
\name{fwaveletspl_3}
\alias{fwaveletspl_3}
\title{Wavelet Transform}
\usage{
fwaveletspl_3(ra0, sjm, det = F)
}
\arguments{
\item{ra0}{A numeric vector representing the input signal.}

\item{sjm}{An integer specifying the number of scales.}

\item{det}{A logical value indicating whether to return the detail coefficients (TRUE) or the approximation coefficients (FALSE). Default is FALSE.}
}
\value{
A matrix containing the wavelet coefficients.
}
\description{
Computes the wavelet transform of a signal using the specified filters and scales.
}
\examples{
# Example usage:
ra0 <- rnorm(1024)  # Example signal
sjm <- 5  # Number of scales
wavelets <- fwaveletspl_3(ra0, sjm)
print(wavelets)
}
