% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{plot.css}
\alias{plot.css}
\title{Plot Curvature Scale Space Image}
\usage{
\method{plot}{css}(x, ...)
}
\arguments{
\item{x}{An object of class `css`.}

\item{...}{Additional arguments passed to plot.}
}
\description{
Visualizes the CSS image, showing inflection points across different scales.
}
