% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_images.R
\name{process_images}
\alias{process_images}
\title{Process Images in Folder}
\usage{
process_images(
  folder,
  subfolder = FALSE,
  threshold = NULL,
  wavelets = TRUE,
  ef = TRUE,
  testing = TRUE,
  pseudolandmarks = "both",
  save = TRUE,
  pixels_per_mm = NULL,
  detect_scale = FALSE
)
}
\arguments{
\item{folder}{A string specifying the path to the folder containing images.}

\item{subfolder}{A logical value indicating whether the folder has subfolders.}

\item{threshold}{A numeric value for binarization. If not provided, the Otsu method is used.}

\item{wavelets}{A logical value indicating whether to obtain wavelets. Default is TRUE.}

\item{ef}{A logical value indicating whether to obtain elliptic Fourier descriptors. Default is TRUE.}

\item{testing}{A logical value indicating whether to save test images. Default is TRUE.}

\item{pseudolandmarks}{A string specifying the type of pseudolandmarks. Options are 'curvilinear', 'polar', 'both'. Default is 'both'.}

\item{save}{A logical value indicating whether to save the results as CSV files. Default is TRUE.}

\item{pixels_per_mm}{A numeric value specifying the scale (pixels per mm). Default is NULL.}

\item{detect_scale}{A logical value indicating whether to automatically detect the scale bar (1mm) in the image. Default is FALSE.}
}
\description{
This function processes images in a specified folder, applying various transformations and analyses.
}
